requirements.txt
anyio==3.6.2
asgiref==3.8.1
asttokens==3.0.0
click==8.1.3
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
Django==5.1.5
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.95.0
h11==0.14.0
httptools==0.5.0
idna==3.4
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pydantic==1.10.7
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pywin32==308
PyYAML==6.0
pyzmq==26.2.0
six==1.17.0
sniffio==1.3.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.26.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.21.1
watchfiles==0.19.0
wcwidth==0.2.13
websockets==11.0.1
